--- a/app/Actions/Commands/ExportCommand.php
+++ b/app/Actions/Commands/ExportCommand.php
@@ -9,4 +9,3 @@ class ExportCommand
         //
     }
 }
-