--- a/app/Actions/Commands/FragCommand.php
+++ b/app/Actions/Commands/FragCommand.php
@@ -7,7 +7,6 @@ use App\DTOs\CommandRequest;
 use App\DTOs\CommandResponse;
 use App\Jobs\ProcessFragmentJob;
 use App\Models\Fragment;
-use Illuminate\Support\Str;
 
 class FragCommand implements HandlesCommand
 {
@@ -19,7 +18,7 @@ class FragCommand implements HandlesCommand
             return new CommandResponse(
                 type: 'frag',
                 shouldShowErrorToast: true,
-                message: "No valid fragment detected. Please try `/frag Your message here...`",
+                message: 'No valid fragment detected. Please try `/frag Your message here...`',
             );
         }